+++ forward_engineering/api.js
@@ -1,36 +1,9 @@
-const _ = require('lodash');
-const { generateVertices, generateEdges } = require('./gremlinHelper');
 const reverseEngineeringApi = require('../reverse_engineering/api');
 const applyToInstanceHelper = require('./applyToInstanceHelper');
+const { generateContainerScript } = require('./generateContainerScript');
 
 module.exports = {
-	generateContainerScript(data, logger, cb) {
-		let { collections, relationships, jsonData } = data;
-		logger.clear();
-		try {
-			let resultScript = '';
-
-			collections = collections.map(JSON.parse);
-			relationships = relationships.map(JSON.parse);
-
-			const verticesScript = generateVertices(collections, jsonData);
-			const edgesScript = generateEdges(collections, relationships, jsonData);
-
-			if (verticesScript) {
-				resultScript += verticesScript;
-			}
-
-			if (edgesScript) {
-				resultScript += '\n\n' + edgesScript;
-			}
-
-			cb(null, resultScript);
-		} catch (e) {
-			logger.log('error', { message: e.message, stack: e.stack }, 'Forward-Engineering Error');
-
-			cb({ message: e.message, stack: e.stack });
-		}
-	},
+	generateContainerScript,
 
 	async applyToInstance(data, logger, cb, app) {
 		try {
